requirements.txt
# Core ML & Data Science
numpy
pandas
scikit-learn
xgboost

# Experiment Tracking
mlflow

# Version Control for Data and Models
dvc

# AWS CLI (for cloud integration)
awscli

# Optional: Logging, Serialization, Utility
joblib
PyYAML

# CLI Enhancers (optional but nice for logs/scripts)
tqdm
click

# If you're using notebooks
jupyter

# If you want linting in CI/CD
flake8

# For unit testing
pytest

nltk 

dagshub 

python-dotenv